d-api-management/src/lists_api/requirements.txt
fastapi
uvicorn
pydantic
azure-cosmos
azure-identity
python-dotenv
opentelemetry-instrumentation-fastapi
opentelemetry-api
opentelemetry-sdk
azure-core-tracing-opentelemetry
azure-monitor-opentelemetry